+++ Docs/changelog.md
@@ -16,12 +16,16 @@ tags: ''
 * New: Mermaid diagram rendering supported via mmdr
 * New: Nomnoml diagram rendering supported via nomnoml
 * New: PlantUML diagram rendering supported. Note: Requires seperate PlantUMl download and installation
+* New: convert diagram2svg command to convert nomnoml and mermaid diagrams to SVG format
 * New: Search command to search text in bookgen project contents
 * New: Support for openCli specification
 * New: Mixcloud media embeding supported
 * New: Html2OpenXml command to convert HTML to OpenXML format for Word processing
 * New: Template engine reworked, now supports functions and is more robust
 * New: Preview command to preview the current project files in a browser
+* New: table-csv syntax highlighting, that renders the data as a table in HTML output
+* New: table-tsv syntax highlighting, that renders the data as a table in HTML output
+* New: table-ssv syntax highlighting, that renders the data as a table in HTML output
 * Change: Commands documentation generated from code and is now more complete
 * Change: App settings reworked
 * Change: Dropped mathjax as dependency for formula rendering. New formula rendering is based on katex and works with markdown syntax